Déplace le calcul de niveau de sécurité dans le moteur V2
