+++ src/moteurRegles/v2/niveauSecurite.ts
@@ -1,0 +1,103 @@
+import {
+  AudienceCible,
+  CategorieDonneesTraitees,
+  DureeDysfonctionnementAcceptable,
+  NiveauSecurite,
+  OuvertureSysteme,
+  questionsV2,
+  VolumetrieDonneesTraitees,
+} from '../../../donneesReferentielMesuresV2.js';
+
+export type NiveauCriticite = 1 | 2 | 3 | 4;
+export type NiveauExposition = 1 | 2 | 3 | 4;
+
+export const matriceCriticiteVolumetrieDonneesTraitees: NiveauCriticite[][] = [
+  [1, 1, 2, 2],
+  [1, 1, 2, 3],
+  [1, 2, 3, 4],
+  [1, 2, 4, 4],
+];
+
+export const matriceCriticiteAudienceCibleDisponibilite: NiveauCriticite[][] = [
+  [1, 1, 2, 3],
+  [1, 1, 2, 3],
+  [1, 1, 3, 4],
+  [1, 1, 3, 4],
+];
+
+export const criticiteVolumetrieDonneesTraitees = (
+  volumetrie: VolumetrieDonneesTraitees,
+  categories: CategorieDonneesTraitees[],
+  autresDonneesTraitees: string[]
+): NiveauCriticite => {
+  const criticiteVolumetrie =
+    questionsV2.volumetrieDonneesTraitees[volumetrie].criticite;
+
+  const criticiteDonnees = categories.map(
+    (c) => questionsV2.categorieDonneesTraitees[c].criticite
+  );
+  if (autresDonneesTraitees.length > 0) {
+    criticiteDonnees.push(1);
+  }
+  return criticiteDonnees
+    .map(
+      (criticite) =>
+        matriceCriticiteVolumetrieDonneesTraitees[criticiteVolumetrie - 1][
+          criticite - 1
+        ]
+    )
+    .reduce((c1, c2) => Math.max(c1, c2) as NiveauCriticite, 1);
+};
+
+export const criticiteDisponibiliteEtAudienceCible = (
+  disponibilite: DureeDysfonctionnementAcceptable,
+  audienceCible: AudienceCible
+): NiveauCriticite => {
+  const criticiteAudienceCible =
+    questionsV2.audienceCible[audienceCible].criticite;
+  const criticiteDisponibilite =
+    questionsV2.dureeDysfonctionnementAcceptable[disponibilite].criticite;
+  return matriceCriticiteAudienceCibleDisponibilite[criticiteAudienceCible - 1][
+    criticiteDisponibilite - 1
+  ];
+};
+
+export const matriceExposition: NiveauExposition[][] = [[1, 2, 2, 3]];
+
+export const niveauExposition = (
+  ouvertureSysteme: OuvertureSysteme
+): NiveauExposition =>
+  matriceExposition[0][
+    questionsV2.ouvertureSysteme[ouvertureSysteme].criticite - 1
+  ];
+
+export const matriceBesoinsSecuriteCriticiteExposition: NiveauSecurite[][] = [
+  ['niveau1', 'niveau1', 'niveau1', 'niveau2'],
+  ['niveau1', 'niveau1', 'niveau2', 'niveau2'],
+  ['niveau2', 'niveau2', 'niveau3', 'niveau3'],
+  ['niveau2', 'niveau3', 'niveau3', 'niveau3'],
+];
+
+export const niveauSecuriteRequis = (
+  volumetrie: VolumetrieDonneesTraitees,
+  categories: CategorieDonneesTraitees[],
+  autresDonneesTraitees: string[],
+  disponibilite: DureeDysfonctionnementAcceptable,
+  audienceCible: AudienceCible,
+  ouvertureSysteme: OuvertureSysteme
+): NiveauSecurite => {
+  const criticiteDonnees = criticiteVolumetrieDonneesTraitees(
+    volumetrie,
+    categories,
+    autresDonneesTraitees
+  );
+  const criticiteDisponibilite = criticiteDisponibiliteEtAudienceCible(
+    disponibilite,
+    audienceCible
+  );
+  const criticiteGenerale = Math.max(criticiteDonnees, criticiteDisponibilite);
+  const exposition = niveauExposition(ouvertureSysteme);
+  return matriceBesoinsSecuriteCriticiteExposition[criticiteGenerale - 1][
+    exposition - 1
+  ];
+};
